[M03] 修复 pipeline.expire 参数类型（long→int）
- Jedis 3.8.0 Pipeline.expire 接受 int，去掉 (long) 强转
- ttlHours*3600 在合理取值范围内不会溢出 int

Refs: DD-V1.2 第10章

+++ cs-copilot-service/src/main/java/com/cmbchina/cs/assitsvc/asr/DialogHistoryManagerImpl.java
@@ -55,7 +55,7 @@ public class DialogHistoryManagerImpl implements DialogHistoryManager {
             pipeline.rpush(key, json);
             // 保留尾部最新的 maxSize 条，超出时裁剪头部旧数据
             pipeline.ltrim(key, -props.getMaxSize(), -1);
-            pipeline.expire(key, (long) props.getTtlHours() * 3600);
+            pipeline.expire(key, props.getTtlHours() * 3600);
             pipeline.sync();
         } catch (JedisException e) {
             log.warn("[M03] Redis append failed, callId={}", callId, e);
